// src/models/BudgetsModel.ts
// export interface BudgetsModel {
//     budgets: {
//         id: string;
//         name: string;
//         createdAt: Date;
//         amount: number;
//         color: string;
//     };
// }

export interface BudgetsModel {
    id: string;
    name: string;
    createdAt: number;
    amount: number;
    color: string;
}